requirement.txt
pytorch
numpy
openCV
matplotlib